SwaggerHelper/requirements.txt
requests
python-dotenv
bs4
playwright